refactoring the LL class
