--- demo/src/main/java/com/karumanchi/linkedlist/ListNode.java
@@ -1,18 +1,0 @@
-package com.karumanchi.linkedlist;
-
-public class ListNode {
-    int val;
-    ListNode next;
-
-    ListNode() {
-    }
-
-    ListNode(int val) {
-        this.val = val;
-    }
-
-    ListNode(int val, ListNode next) {
-        this.val = val;
-        this.next = next;
-    }
-}
