+++ demo/src/main/java/com/karumanchi/linkedlist/helper/ListNode.java
@@ -1,0 +1,19 @@
+package com.karumanchi.linkedlist.helper;
+
+public class ListNode {
+    
+    public int val;
+    public ListNode next;
+
+    public ListNode() {
+    }
+
+    public ListNode(int val) {
+        this.val = val;
+    }
+
+    public ListNode(int val, ListNode next) {
+        this.val = val;
+        this.next = next;
+    }
+}
