+++ demo/src/main/java/com/karumanchi/linkedlist/helper/TraverseLinkedList.java
@@ -1,0 +1,21 @@
+package com.karumanchi.linkedlist.helper;
+
+public class TraverseLinkedList {
+    
+    public static void traverse(Node head) {
+        Node node = head;
+        while(node != null) {
+            System.out.println(node.getData());
+            node = node.getNext();
+        }
+    }
+
+    public static void traverseLL(ListNode head) {
+        ListNode node = head;
+        while(node != null) {
+            System.out.println(node.val);
+            node = node.next;
+        }
+    }
+    
+}
